+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## [Unreleased]
+
+### Added
+- Added a "Refresh" button to update the panel content.
+- add the logic for displaying unused packages and files.
+- Added an "Analyze Project" button to run a static analyzer on the project.
+- Display project name in the toolbar.
+- Improved styling for buttons and results display.
+
+### Changed
+- move the open panel button from pubspec.yaml to the toolbar.
+
+### Fixed
+- Fix #1
+
 ## [0.0.4]
 
 - First release of Pub Manager.